LayoutTests/imported/w3c/web-platform-tests/html/semantics/document-metadata/the-meta-element/pragma-directives/attr-meta-http-equiv-refresh/allow-scripts-flag-changing-2-expected.txt

CONSOLE MESSAGE: Unable to do meta refresh due to sandboxing

PASS Meta refresh of the original iframe is not blocked if moved into a sandboxed iframe
